Pong Unity & english Update U 2021.3.6.f1
this is my first unity project, developed in april 2021.
Is the clasic Pong but only have local multyplayer.
This game helped to improve and practice my use of rotations, translations, collisions, physics, physical materials and counters on a canvas.
It has an extremely simple problematic design, but I try to add a personal touch to it